rptest: Break out ScaleParameters class
